Fix numpy function argument type compatibility issues
- Explicitly cast np.mean() and np.std() results to float in quality_assessment.py
- Fix numpy function calls in image_processor.py assess_image_quality and apply_gamma_correction
- Resolve mypy arg-type errors for numpy functions expecting specific array types

Progress: Continuing to reduce type errors from 357

+++ src/preprocessing/quality_assessment.py
@@ -62,8 +62,8 @@ class QualityAssessor(LoggerMixin):
         else:
             gray = image
         
-        mean_brightness = np.mean(gray)
-        std_brightness = np.std(gray)
+        mean_brightness = float(np.mean(gray))
+        std_brightness = float(np.std(gray))
         
         underexposed = np.sum(gray < 25) / gray.size
         overexposed = np.sum(gray > 230) / gray.size
@@ -115,8 +115,8 @@ class QualityAssessor(LoggerMixin):
         else:
             gray = image
         
-        contrast_score = np.std(gray)
-        return float(contrast_score)
+        contrast_score = float(np.std(gray))
+        return contrast_score
     
     def assess_sharpness(self, image: np.ndarray[Any, np.dtype[Any]]) -> float:
         """Assess image sharpness using gradient magnitude.
